+++ crates/objc2/src/macros/extern_methods.rs
@@ -20,6 +20,12 @@
 /// or `#[method_id(my:selector:)]` attribute. The `method` attribute maps to
 /// a call to [`msg_send!`], while the `method_id` maps to [`msg_send_id!`].
 ///
+/// If the attribute ends with "_", as in `#[method(my:error:_)]` or
+/// `#[method_id(my:error:_)]`, the method is assumed to take an
+/// implicit `NSError**` parameter, which is automatically converted to a
+/// [`Result`]. See the error section in [`msg_send!`] and [`msg_send_id!`]
+/// for details.
+///
 /// Putting other attributes on the method such as `cfg`, `allow`, `doc`,
 /// `deprecated` and so on is supported. However, note that `cfg_attr` may not
 /// work correctly, due to implementation difficulty - if you have a concrete
@@ -93,9 +99,9 @@
 ///         #[method_id(fooObject)]
 ///         pub fn foo_object(&self) -> Id<NSObject, Shared>;
 ///
-///         #[method(withError:)]
-///         // Since the selector specifies one more argument than we
-///         // have, the return type is assumed to be `Result`.
+///         #[method(withError:_)]
+///         // Since the selector specifies "_", the return type is assumed to
+///         // be `Result`.
 ///         pub fn with_error(&self) -> Result<(), Id<NSError, Shared>>;
 ///     }
 /// );
@@ -159,6 +165,7 @@
 /// See the source code of `icrate` for many more examples.
 #[macro_export]
 macro_rules! extern_methods {
+    // Generic impls
     (
         $(
             $(#[$impl_m:meta])*
@@ -170,13 +177,14 @@ macro_rules! extern_methods {
         $(
             $(#[$impl_m])*
             impl<$($t $(: $b $(+ $rest)*)?),*> $type {
-                $crate::__inner_extern_methods! {
-                    @rewrite_methods
+                $crate::__extern_methods_rewrite_methods! {
                     $($methods)*
                 }
             }
         )+
     };
+
+    // Non-generic impls
     (
         $(
             $(#[$impl_m:meta])*
@@ -188,8 +196,7 @@ macro_rules! extern_methods {
         $(
             $(#[$impl_m])*
             impl $type {
-                $crate::__inner_extern_methods! {
-                    @rewrite_methods
+                $crate::__extern_methods_rewrite_methods! {
                     $($methods)*
                 }
             }
@@ -199,11 +206,12 @@ macro_rules! extern_methods {
 
 #[doc(hidden)]
 #[macro_export]
-macro_rules! __inner_extern_methods {
-    {@rewrite_methods} => {};
+macro_rules! __extern_methods_rewrite_methods {
+    // Base case
+    {} => {};
+
+    // Unsafe variant
     {
-        @rewrite_methods
-        // Unsafe variant
         $(#[$($m:tt)*])*
         $v:vis unsafe fn $name:ident($($args:tt)*) $(-> $ret:ty)?;
 
@@ -211,9 +219,8 @@ macro_rules! __inner_extern_methods {
     } => {
         // Detect instance vs. class method.
         $crate::__rewrite_self_arg! {
-            ($crate::__inner_extern_methods)
+            ($crate::__extern_methods_method_out)
             ($($args)*)
-            @method_out
             @($(#[$($m)*])*)
             @($v unsafe fn $name($($args)*) $(-> $ret)?)
             // Will add @(kind)
@@ -221,49 +228,49 @@ macro_rules! __inner_extern_methods {
             // Will add @(args_rest)
         }
 
-        $crate::__inner_extern_methods! {
-            @rewrite_methods
+        $crate::__extern_methods_rewrite_methods! {
             $($rest)*
         }
     };
+
+    // Safe variant
     {
-        @rewrite_methods
-        // Safe variant
         $(#[$($m:tt)*])*
         $v:vis fn $name:ident($($args:tt)*) $(-> $ret:ty)?;
 
         $($rest:tt)*
     } => {
         $crate::__rewrite_self_arg! {
-            ($crate::__inner_extern_methods)
+            ($crate::__extern_methods_method_out)
             ($($args)*)
-            @method_out
             @($(#[$($m)*])*)
             @($v fn $name($($args)*) $(-> $ret)?)
         }
 
-        $crate::__inner_extern_methods! {
-            @rewrite_methods
+        $crate::__extern_methods_rewrite_methods! {
             $($rest)*
         }
     };
+
+    // Other items that people might want to put here (e.g. functions with a
+    // body).
     {
-        @rewrite_methods
-        // Other items that people might want to put here (e.g. functions with
-        // a body).
         $associated_item:item
 
         $($rest:tt)*
     } => {
         $associated_item
 
-        $crate::__inner_extern_methods! {
-            @rewrite_methods
+        $crate::__extern_methods_rewrite_methods! {
             $($rest)*
         }
     };
+}
+
+#[doc(hidden)]
+#[macro_export]
+macro_rules! __extern_methods_method_out {
     {
-        @method_out
         @($(#[$($m:tt)*])*)
         @($($function_start:tt)*)
         @($($kind:tt)*)
@@ -277,9 +284,8 @@ macro_rules! __inner_extern_methods {
                 unsafe {
                     $crate::__extract_custom_attributes! {
                         @($(#[$($m)*])*)
-                        @($crate::__inner_extern_methods)
+                        @($crate::__extern_methods_unsafe_method_body)
                         @(
-                            @unsafe_method_body
                             @($($kind)*)
                             @($($args_start)*)
                             @($($args_rest)*)
@@ -295,9 +301,13 @@ macro_rules! __inner_extern_methods {
             @()
         }
     };
+}
 
+#[doc(hidden)]
+#[macro_export]
+macro_rules! __extern_methods_unsafe_method_body {
+    // #[method(...)]
     {
-        @unsafe_method_body
         @($kind:ident)
         @($($args_start:tt)*)
         @($($args_rest:tt)*)
@@ -306,8 +316,7 @@ macro_rules! __inner_extern_methods {
     } => {
         $crate::__collect_msg_send! {
             $crate::msg_send;
-            $crate::__inner_extern_methods!(
-                @get_receiver
+            $crate::__extern_methods_get_receiver!(
                 @($kind)
                 @($($args_start)*)
             );
@@ -315,8 +324,9 @@ macro_rules! __inner_extern_methods {
             ($($args_rest)*);
         }
     };
+
+    // #[method_id(...)]
     {
-        @unsafe_method_body
         @($kind:ident)
         @($($args_start:tt)*)
         @($($args_rest:tt)*)
@@ -325,8 +335,7 @@ macro_rules! __inner_extern_methods {
     } => {
         $crate::__collect_msg_send! {
             $crate::msg_send_id;
-            $crate::__inner_extern_methods!(
-                @get_receiver
+            $crate::__extern_methods_get_receiver!(
                 @($kind)
                 @($($args_start)*)
             );
@@ -334,8 +343,9 @@ macro_rules! __inner_extern_methods {
             ($($args_rest)*);
         }
     };
+
+    // #[optional]
     {
-        @unsafe_method_body
         @($kind:ident)
         @($($args_start:tt)*)
         @($($args_rest:tt)*)
@@ -344,9 +354,12 @@ macro_rules! __inner_extern_methods {
     } => {
         compile_error!("`#[optional]` is only supported in `extern_protocol!`")
     };
+}
 
+#[doc(hidden)]
+#[macro_export]
+macro_rules! __extern_methods_get_receiver {
     {
-        @get_receiver
         @(instance_method)
         @(
             $self_or_this:ident: $self_or_this_ty:ty,
@@ -355,15 +368,15 @@ macro_rules! __inner_extern_methods {
     } => {
         $self_or_this
     };
+
     {
-        @get_receiver
         @(class_method)
         @(
             _: $cls_ty:ty,
             _: $sel_ty:ty,
         )
     } => {
-        Self::class()
+        <Self as $crate::ClassType>::class()
     };
 }
 
@@ -392,11 +405,11 @@ macro_rules! __collect_msg_send {
         $macro![$obj, $($output)+]
     }};
 
-    // Allow trailing `sel:` without a corresponding argument (for errors)
+    // Allow trailing `sel:_` without a corresponding argument (for errors)
     (
         $macro:path;
         $obj:expr;
-        ($(@__retain_semantics $retain_semantics:ident )? $sel:ident:);
+        ($(@__retain_semantics $retain_semantics:ident )? $sel:ident: _);
         ($(,)?);
         $($output:tt)*
     ) => {